+++ src/main/java/com/imveis/visita/Imoveis/dtos/AgendamentoRequest.java
@@ -2,14 +2,13 @@ package com.imveis.visita.Imoveis.dtos;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 @Data
 public class AgendamentoRequest {
-    private BigInteger imovelId;
+    private Long imovelId;
     private String nomeVisitante;
     private LocalDate dataAgendamento;
-    private BigInteger usuarioId;
+    private Long usuarioId;
     private Boolean horarioMarcado;
 }
