+++ src/main/java/com/imveis/visita/Imoveis/dtos/CorretorRequest.java
@@ -2,11 +2,9 @@ package com.imveis.visita.Imoveis.dtos;
 
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 public class CorretorRequest {
-    private BigInteger usuarioId;
+    private Long usuarioId;
     private String creci;
 }
