+++ src/main/java/com/imveis/visita/Imoveis/dtos/PropostaRequest.java
@@ -3,7 +3,6 @@ package com.imveis.visita.Imoveis.dtos;
 import lombok.Data;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 
 @Data
 public class PropostaRequest {
@@ -11,5 +10,5 @@ public class PropostaRequest {
     private BigDecimal rendaMensal;
     private BigDecimal entrada;
     private BigDecimal valorImovel;
-    private BigInteger idImovel;
+    private Long idImovel;
 }
