+++ src/main/java/com/imveis/visita/Imoveis/dtos/RelatorioAgendamentoDTO.java
@@ -4,14 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
-
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class RelatorioAgendamentoDTO {
-    private BigInteger idImovel;
+    private Long idImovel;
     private String descricaoImovel;
     private long quantidadeAgendamentos;
 }
