+++ src/main/java/com/imveis/visita/Imoveis/dtos/RelatorioUsuarioDTO.java
@@ -4,13 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class RelatorioUsuarioDTO {
-    private BigInteger idUsuario;
+    private Long idUsuario;
     private String nomeUsuario;
     private long quantidadeAcessos;
 }
