+++ src/main/java/com/imveis/visita/Imoveis/dtos/RelatorioVistoriaDTO.java
@@ -4,20 +4,19 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 @Data
 @Getter
 @Setter
 public class RelatorioVistoriaDTO {
-    private BigInteger idVistoria;
-    private BigInteger idImovel;
+    private Long idVistoria;
+    private Long idImovel;
     private String descricaoImovel;
     private LocalDate dataVistoria;
     private String laudoVistoria;
 
-    public RelatorioVistoriaDTO(BigInteger idVistoria, BigInteger idImovel, String descricaoImovel, LocalDate dataVistoria, String laudoVistoria) {
+    public RelatorioVistoriaDTO(Long idVistoria, Long idImovel, String descricaoImovel, LocalDate dataVistoria, String laudoVistoria) {
         this.idVistoria = idVistoria;
         this.idImovel = idImovel;
         this.descricaoImovel = descricaoImovel;
