+++ src/main/java/com/imveis/visita/Imoveis/dtos/VistoriaRequest.java
@@ -4,18 +4,17 @@ import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
 import java.time.LocalDate;
 
 @Getter
 @Setter
 @Data
 public class VistoriaRequest {
-    private BigInteger idVistoria;
+    private Long idVistoria;
     private String tipoVistoria;
     private String laudoVistoria;
     private LocalDate dataVistoria;
-    private BigInteger usuarioId;
+    private Long usuarioId;
     private String rua;
     private String numero;
     private String bairro;
