+++ src/main/java/com/imveis/visita/Imoveis/repositories/EnderecoRepository.java
@@ -4,8 +4,7 @@ import com.imveis.visita.Imoveis.entities.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 
 @Repository
-public interface EnderecoRepository extends JpaRepository<Endereco, BigInteger> {
+public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
 }
