+++ src/main/java/com/imveis/visita/Imoveis/repositories/FotoImovelRepository.java
@@ -4,9 +4,8 @@ import com.imveis.visita.Imoveis.entities.FotoImovel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 
 @Repository
-public interface FotoImovelRepository extends JpaRepository<FotoImovel, BigInteger>{
+public interface FotoImovelRepository extends JpaRepository<FotoImovel, Long> {
 
 }
