+++ src/main/java/com/imveis/visita/Imoveis/repositories/FotoVistoriaRepository.java
@@ -4,10 +4,9 @@ import com.imveis.visita.Imoveis.entities.FotoVistoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 
 @Repository
-public interface FotoVistoriaRepository extends JpaRepository<FotoVistoria, BigInteger> {
+public interface FotoVistoriaRepository extends JpaRepository<FotoVistoria, Long> {
 }
