+++ src/main/java/com/imveis/visita/Imoveis/repositories/FuncionarioRepository.java
@@ -4,11 +4,10 @@ import com.imveis.visita.Imoveis.entities.Funcionario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface FuncionarioRepository extends JpaRepository<Funcionario, BigInteger> {
+public interface FuncionarioRepository extends JpaRepository<Funcionario, Long> {
     @SuppressWarnings("unused")
     Optional<Funcionario> findByEmailAndSenha(String email, String senha);
 }
