+++ src/main/java/com/imveis/visita/Imoveis/repositories/ImobiliariaRepository.java
@@ -4,10 +4,9 @@ import com.imveis.visita.Imoveis.entities.Imobiliaria;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface ImobiliariaRepository extends JpaRepository<Imobiliaria, BigInteger>{
+public interface ImobiliariaRepository extends JpaRepository<Imobiliaria, Long> {
     Optional<Imobiliaria> findByCnpj(String cnpj);
 }
