+++ src/main/java/com/imveis/visita/Imoveis/repositories/RoleRepository.java
@@ -4,11 +4,10 @@ import com.imveis.visita.Imoveis.entities.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, BigInteger> {
+public interface RoleRepository extends JpaRepository<Role, Long> {
     Optional<Role> findByNome(String nome);
 
     boolean existsByNome(String nome);
